+++ nuxt-sand/components/daisyUISample/organisms/Navigation.vue
@@ -2,7 +2,7 @@
     <header class="bg-white">
       <nav class="mx-auto flex max-w-7xl items-center justify-between p-6 lg:px-8" aria-label="Global">
         <div class="flex lg:flex-1">
-          <a href="/" class="-m-1.5 p-1.5">
+          <a href="/sampleTop" class="-m-1.5 p-1.5">
             <span class="sr-only">Your Company</span>
             <img class="h-8 w-auto" src="https://tailwindui.com/img/logos/mark.svg?color=indigo&shade=600" alt="" />
           </a>
