+++ nuxt-sand/layouts/README.md
@@ -1,0 +1,6 @@
+## layout dir
+共通のレイアウトを定義するディレクトリ。
+
+
+## reference
+- [公式ドキュメント](https://nuxt.com/docs/guide/directory-structure/layouts)
