+++ nuxt-sand/layouts/default.vue
@@ -1,0 +1,10 @@
+
+import { DaisyUISampleOrganismsNavigation } from '~~/.nuxt/components';
+
+<template>
+    <div>
+        <CommonOrganismsNavigation />
+      <slot />
+    </div>
+  </template>
+
